fix: Added option to remove errors on non-macOS
The 'IgnoreUnknown' option was added to the SSH config with the value
'UseKeychain,AddKeysToAgent' as these options are for macOS only, using
this option tells SSH to ignore the listed settings if it does not
understand them.

This should stop "Unknown Option: UseKeychain" on linux hosts and make
the config file more portable.